Remove `build.py` by distributing it's content close to the source code (optimize) or in the cache utilities
