--- docs/docstrings/dsl/dace/build.md
@@ -1,12 +1,0 @@
-# build
-
-::: dsl.dace.build
-
-<style>
-/* re-enable the left side navigation bar for this page */
-@media screen and (min-width: 76.1875em) {
-  .md-sidebar--primary {
-    display: block !important;
-  }
-}
-</style>
